+++ tests/testthat/test-hardhat_interfaces.R
@@ -11,6 +11,9 @@ test_that("Training regression for data.frame and formula", {
   expect_no_error(
     predict(fit, x)
   )
+  expect_no_error(
+    augment(fit, ames)
+  )
 
   expect_no_error(
     fit <- tabnet_fit(x, y, epochs = 2, verbose = TRUE)
@@ -29,6 +32,14 @@ test_that("Training classification for data.frame", {
 
   expect_no_error(
     predict(fit, attrix)
+  )
+
+  expect_no_error(
+    augment(fit, attrition[ids,] , type = "prob")
+  )
+
+  expect_no_error(
+    augment(fit, attrition[ids,] )
   )
 
 })
@@ -103,6 +114,7 @@ test_that("checkpoints works for inference", {
 
   expect_equal(p3, predict(fit, x))
 
+  expect_no_error(augment(fit, ames[ids,], epoch = 1))
 })
 
 test_that("print module works even after a reload from disk", {